requirements.txt
flask
google-genai
pillow
gunicorn